test: unify fixtures in conftest
