+++ tests/test_dataingestor.py
@@ -13,74 +13,21 @@ from verdesat.ingestion.dataingestor import DataIngestor
 from verdesat.geo.aoi import AOI
 
 
-@pytest.fixture
-def dummy_aoi():
-    """
-    Create a dummy AOI with a simple square Polygon and an 'id' in static_props.
-    """
-    geom = Polygon([(0, 0), (0, 1), (1, 1), (1, 0)])
-    return AOI(geometry=geom, static_props={"id": 1})
-
-
-@pytest.fixture
-def dummy_sensor():
-    """
-    Dummy sensor with minimal attributes and a pass‑through compute_index.
-    """
-
-    class DummySensor:
-        def __init__(self):
-            self.collection_id = "dummy/collection"
-
-        @staticmethod  # noqa: D401
-        def compute_index(img, index):  # pylint: disable=unused-argument
-            # Return a dummy image that has a 'reduceRegions' method in get_image_collection
-            return img
-
-    return DummySensor()
-
-
-@pytest.fixture
-def _dummy_ee_manager(monkeypatch):
-    """
-    Monkeypatch ee_manager.get_image_collection to return a Fake FeatureCollection.
-    whose getInfo() produces a known feature list.
-    """
-
-    # Create a fake FeatureCollection class
-    class FakeFC:
-        """Stub of an Earth Engine FeatureCollection for testing."""
-
-        def __init__(self, features):
-            self._features = features
-
-        def map(self, func):  # pylint: disable=unused-argument
-            return self
-
-        def flatten(self):
-            return self
-
-        def getInfo(self):  # pylint: disable=invalid-name
-            # Simulate EarthEngine getInfo output: list of features with properties
-            return {
-                "features": [
-                    {"properties": {"id": 1, "date": "2020-01-01", "mean": 0.5}}
-                ]
-            }
-
-    # Monkeypatch the ee_manager used in DataIngestor
-    monkeypatch.setattr(
-        "verdesat.ingestion.dataingestor.ee_manager.get_image_collection",
-        lambda *args, **kwargs: FakeFC([None]),
-    )
-    return FakeFC
-
-
-def test_download_timeseries_no_aggregation(dummy_aoi, dummy_sensor, _dummy_ee_manager):
+def test_download_timeseries_no_aggregation(
+    dummy_aoi, dummy_sensor, _dummy_ee_manager, monkeypatch
+):
     """
     Test that DataIngestor.download_timeseries returns a DataFrame with correct columns
     for one daily chunk, without aggregation (freq=None).
     """
+    # Stub out chunked retrieval to return a known DataFrame
+    monkeypatch.setattr(
+        DataIngestor,
+        "_chunked_timeseries",
+        lambda self, aoi, start_date, end_date, scale, index, chunk_freq: pd.DataFrame(
+            [{"id": 1, "date": "2020-01-01", f"mean_{index}": 0.5}]
+        ),
+    )
     di = DataIngestor(sensor=dummy_sensor)
     df = di.download_timeseries(
         dummy_aoi,
@@ -101,12 +48,24 @@ def test_download_timeseries_no_aggregation(dummy_aoi, dummy_sensor, _dummy_ee_m
 
 
 def test_download_timeseries_with_aggregation(
-    dummy_aoi, dummy_sensor, _dummy_ee_manager
+    dummy_aoi, dummy_sensor, _dummy_ee_manager, monkeypatch
 ):
     """
     Test that DataIngestor.download_timeseries with freq='M' returns a DataFrame
     resampled to monthly (here, single month so no change).
     """
+    # Stub out chunked retrieval to return a known daily DataFrame over January 2020
+    raw_data = [
+        {"id": 1, "date": f"2020-01-{day:02d}", "mean_ndvi": 0.5}
+        for day in range(1, 32)
+    ]
+    monkeypatch.setattr(
+        DataIngestor,
+        "_chunked_timeseries",
+        lambda self, aoi, start_date, end_date, scale, index, chunk_freq: pd.DataFrame(
+            raw_data
+        ),
+    )
     di = DataIngestor(sensor=dummy_sensor)
     df_agg = di.download_timeseries(
         dummy_aoi,
