+++ tests/test_eemanager.py
@@ -12,64 +12,27 @@ from verdesat.ingestion.eemanager import EarthEngineManager, ee_manager
 from verdesat.ingestion.sensorspec import SensorSpec
 
 
-@pytest.fixture(autouse=True)
-def mock_ee(monkeypatch):
-    """
-    Monkeypatch ee.Initialize, ee.ServiceAccountCredentials, and ee.ImageCollection.
-    This ensures EarthEngineManager.initialize() and get_image_collection() can run
-    without real GEE.
-    """
-    # Stub any credential calls
-    monkeypatch.setattr(ee, "Initialize", lambda *args, **kwargs: None)
-    monkeypatch.setattr(ee, "ServiceAccountCredentials", lambda a, b: MagicMock())
 
-    # Provide a dummy ImageCollection that supports .filterDate, .filterBounds, .map
-    class DummyCollection:
-        """
-        Dummy ImageCollection stub that tracks .map() calls.
-        """
+# @pytest.fixture
+# def dummy_sensor(monkeypatch):
+#     """
+#     Create a fake SensorSpec that simply returns the input image when applying cloud_mask.
+#     """
 
-        def __init__(self):
-            self.mapped_with = None
+#     class DummySensor:
+#         """Dummy sensor spec for testing cloud_mask invocation."""
 
-        # pylint: disable=C0103  # filterDate not snake_case
-        def filterDate(self, s, e):
-            """Stub for ImageCollection.filterDate."""
-            return self
+#         def __init__(self):
+#             self.collection_id = "dummy/collection"
 
-        def filterBounds(self, region):
-            """Stub for ImageCollection.filterBounds."""
-            return self
+#         @staticmethod
+#         def cloud_mask(img):
+#             """Apply a no-op cloud mask."""
+#             return img  # No-op mask
 
-        def map(self, func):
-            """Record that .map() was called with the given function."""
-            self.mapped_with = func
-            return self
-
-    monkeypatch.setattr(ee, "ImageCollection", lambda cid: DummyCollection())
-    yield
-
-
-@pytest.fixture
-def dummy_sensor(monkeypatch):
-    """
-    Create a fake SensorSpec that simply returns the input image when applying cloud_mask.
-    """
-
-    class DummySensor:
-        """Dummy sensor spec for testing cloud_mask invocation."""
-
-        def __init__(self):
-            self.collection_id = "dummy/collection"
-
-        @staticmethod
-        def cloud_mask(img):
-            """Apply a no-op cloud mask."""
-            return img  # No-op mask
-
-    # Override SensorSpec.from_collection_id to return DummySensor
-    monkeypatch.setattr(SensorSpec, "from_collection_id", lambda cid: DummySensor())
-    return DummySensor()
+#     # Override SensorSpec.from_collection_id to return DummySensor
+#     monkeypatch.setattr(SensorSpec, "from_collection_id", lambda cid: DummySensor())
+#     return DummySensor()
 
 
 def test_initialize_does_not_raise():
